Modif entités
Suppression Conseil. Créations entités Spot, Topo, Secteur, Voie, Longueur. Ajout des relations pour les entités. Création des repositories manquants (et modif des autres)

+++ src/main/java/com/openclassrooms/escalade/repositories/CommentaireRepository.java
@@ -3,5 +3,5 @@ package com.openclassrooms.escalade.repositories;
 import com.openclassrooms.escalade.entities.Commentaire;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CommentaireRepository extends CrudRepository<Commentaire, Integer> {
+public interface CommentaireRepository extends CrudRepository<Commentaire, Long> {
 }
