+++ src/main/java/com/openclassrooms/escalade/repositories/PhotoRepository.java
@@ -3,5 +3,5 @@ package com.openclassrooms.escalade.repositories;
 import com.openclassrooms.escalade.entities.Photo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PhotoRepository extends CrudRepository<Photo, Integer> {
+public interface PhotoRepository extends CrudRepository<Photo, Long> {
 }
