+++ src/main/java/com/openclassrooms/escalade/repositories/UtilisateurRepository.java
@@ -3,5 +3,5 @@ package com.openclassrooms.escalade.repositories;
 import com.openclassrooms.escalade.entities.Utilisateur;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UtilisateurRepository extends CrudRepository<Utilisateur, Integer> {
+public interface UtilisateurRepository extends CrudRepository<Utilisateur, Long> {
 }
